+++ CHANGELOG.md
@@ -1,13 +1,16 @@
-# Version 0.6.2 (Upcoming, not released yet)
+# Version 0.7.0
 
 ## Bugfixes
 
-* Remove wrong and invalid use statements from ContextConnection class
 * Add Guzzle dependency to composer.json
+* Remove wrong and invalid use statements from ContextConnection class
 
 ## Features
 
 * Add PHPUnit test class for ContextConnection
+* Add LocalContextConnection and LocalConnectionFactory
+* Remove connect(), disconnect() and getSocket() methods from Connection interface
+* Rename ContextConnection to RemoteContextConnection and ConnectionFactory to RemoteConnectionFactory
 
 # Version 0.6.1